+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.12.2](https://github.com/witchesofthehill/manabrew/compare/v1.12.1...v1.12.2) (2026-07-19)
+
+## [1.12.1](https://github.com/witchesofthehill/manabrew/compare/v1.12.0...v1.12.1) (2026-07-17)
+
+## [1.12.0](https://github.com/witchesofthehill/manabrew/compare/v1.11.0...v1.12.0) (2026-07-17)
+
+### Features
+
+* manabrew-protocol v1 ([#445](https://github.com/witchesofthehill/manabrew/issues/445)) ([059d559](https://github.com/witchesofthehill/manabrew/commit/059d5597620cb610b1793f24ebbd7beba7305349))
+
+## [1.11.0](https://github.com/witchesofthehill/manabrew/compare/v1.10.0...v1.11.0) (2026-07-17)
+
+### Features
+
+* **ui:** respect safe-area insets across all screens ([#463](https://github.com/witchesofthehill/manabrew/issues/463)) ([dc5c178](https://github.com/witchesofthehill/manabrew/commit/dc5c178a88d4e13b8d1e4b48bd3b95f450910a69))
+
 ## [1.10.0](https://github.com/witchesofthehill/manabrew/compare/v1.9.4...v1.10.0) (2026-07-15)
 
 ### Features